requirements.txt
python-dotenv
pyaudio
websocket-client==1.6.0
livekit-plugins-openai
livekit-plugins-silero
openai
dotenv
